Changing troop id from byte to ushort. Making it so you need at least 40 upkeep to join an assignment.

+++ Game/Data/Troop/CityTroopStubFactory.cs
@@ -9,7 +9,7 @@
 
         private ICity City { get; set; }
 
-        public TroopStub CreateTroopStub(byte troopId)
+        public TroopStub CreateTroopStub(ushort troopId)
         {
             return new TroopStub(troopId, City);
         }
